Turn off shared sysconfig IOs
Turn off shared sysconfig IOs prior to accessing the fabric over
the SPI bus. This is done using the axonfabric driver via a sysfs
file exported by the driver. This should be done in order to
prevent conflicts on SPI_B_MISO, which is normally driven by the
fabric logic (when programmed and running).